pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quoridor-gaz"
version = "0.1.0"
description = "Quoridor implementation for Gumbel AlphaZero training"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "pyyaml>=6.0",
    "gymnasium>=0.29.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
